+++ lunatrace/bsl/ingest-worker/pkg/metadata/mapper/graphql.go
@@ -12,8 +12,8 @@
 package mapper
 
 import (
-	gql2 "github.com/lunasec-io/lunasec/lunatrace/gen/gql"
-	"github.com/lunasec-io/lunasec/lunatrace/gen/gql/types"
+	gql2 "github.com/lunasec-io/lunasec/lunatrace/gogen/gql"
+	"github.com/lunasec-io/lunasec/lunatrace/gogen/gql/types"
 	"time"
 
 	"github.com/lunasec-io/lunasec/lunatrace/bsl/ingest-worker/pkg/metadata"
