+++ lunatrace/bsl/ingest-worker/pkg/vulnerability/advisory/map.go
@@ -13,7 +13,7 @@ package advisory
 
 import (
 	"errors"
-	"github.com/lunasec-io/lunasec/lunatrace/gen/gql/types"
+	"github.com/lunasec-io/lunasec/lunatrace/gogen/gql/types"
 
 	"strings"
