+++ lunatrace/bsl/ingest-worker/pkg/vulnerability/advisory/osv.go
@@ -17,8 +17,8 @@ import (
 	"github.com/facebookincubator/nvdtools/cvss3"
 	"github.com/lunasec-io/lunasec/lunatrace/bsl/ingest-worker/pkg/vulnerability/schema"
 	"github.com/lunasec-io/lunasec/lunatrace/cli/pkg/util"
-	"github.com/lunasec-io/lunasec/lunatrace/gen/gql"
-	"github.com/lunasec-io/lunasec/lunatrace/gen/gql/types"
+	"github.com/lunasec-io/lunasec/lunatrace/gogen/gql"
+	"github.com/lunasec-io/lunasec/lunatrace/gogen/gql/types"
 	"github.com/rs/zerolog/log"
 	"golang.org/x/exp/slices"
 	"strings"
