+++ lunatrace/bsl/ingest-worker/pkg/vulnerability/advisory/types.go
@@ -12,7 +12,7 @@
 package advisory
 
 import (
-	"github.com/lunasec-io/lunasec/lunatrace/gen/gql"
+	"github.com/lunasec-io/lunasec/lunatrace/gogen/gql"
 )
 
 var (
