+++ lunatrace/bsl/ingest-worker/pkg/vulnerability/ingester.go
@@ -15,7 +15,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/Khan/genqlient/graphql"
-	gql2 "github.com/lunasec-io/lunasec/lunatrace/gen/gql"
+	gql2 "github.com/lunasec-io/lunasec/lunatrace/gogen/gql"
 	"github.com/rs/zerolog/log"
 	"github.com/schollz/progressbar/v3"
 	"github.com/vektah/gqlparser/v2/gqlerror"
