+++ lunatrace/cli/license/upsert_test.go
@@ -16,8 +16,8 @@ package license
 
 import (
 	"context"
-	gql2 "github.com/lunasec-io/lunasec/lunatrace/gen/gql"
-	"github.com/lunasec-io/lunasec/lunatrace/gen/gql/types"
+	gql2 "github.com/lunasec-io/lunasec/lunatrace/gogen/gql"
+	"github.com/lunasec-io/lunasec/lunatrace/gogen/gql/types"
 	"testing"
 	"time"
