+++ lunatrace/gogen/gql/graphql/upsertVulnerabilities.graphql
