+++ lunatrace/gogen/gql/graphql/vulnerability.graphql
